chore: Adjust README example to work w/ copy paste
The example in the README begins with an `import` statement, which would
seem to invite readers to copy and paste the code block in their editor
just to see if everything's working. Since it uses `let` where a
declaration is expected, it doesn't actually work as written.

This just changes the `let`s to `def`s to make it work when copied
and pasted verbatim.